requirements.txt
transformers
datasets
jax[cuda11_cudnn82]>=0.4.2
torch
optax
flax
hydra-core
wandb